♻️move to StringTable
